requirements.txt
discord.py==2.4.0
Pillow==11.0.0